requirements.txt
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.2
beautifulsoup4==4.9.3
requests==2.32.3
python-dotenv==1.0.1